% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_tech_intensity.R
\name{addTechIntensity}
\alias{addTechIntensity}
\title{Join the tech intensity data frame to the yearly WIOD}
\usage{
addTechIntensity(yearly.raw, isic)
}
\arguments{
\item{yearly.raw}{yearly raw data from the downloaded zip such as WIOT2011_October16_ROW.RData}

\item{isic}{parameter to be given for the aggregation of the industries.}
}
\value{
data frame
}
\description{
Join the created technology intensity data frame to the
    yearly raw WIOD wide table so that similar country and technology
    intensity manufacturing industies can be aggregated.
}
