% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_wiod.R
\name{checkWIOD}
\alias{checkWIOD}
\title{Check if the downloaded file is the one downloaded by Altay.}
\usage{
checkWIOD(original.file, download.dir)
}
\arguments{
\item{original.file}{the zip file downloaded.}

\item{download.dir}{directory where the zip file is saved}
}
\description{
Check the integrity of the zip file with the md5sum
    obtained by Altay. The md5sum is not provided by the WIOD source.
}
