% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dom_int_trade.R
\name{domIntTrade}
\alias{domIntTrade}
\title{Generate a long table for domestic and international trade}
\usage{
domIntTrade(net.long)
}
\arguments{
\item{net.long}{yearly long table of source target weight network data}
}
\value{
data frame
}
\description{
Generate yearly long tables for final and itermediate,
    in and out, domestic and internatioanl trade values for each
    economic sectors.
}
