% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_wiod.R
\name{downloadWIOD}
\alias{downloadWIOD}
\title{Download the WIOD zip file from its web site.}
\usage{
downloadWIOD(download.dir)
}
\arguments{
\item{download.dir}{directory to download the original WIOD data.}
}
\description{
Download WIOD zip file from the project web page. The
    downloaded file is the 2016 version which contains RData for each
    year for the period 2000:2014. The integrity of the downloaded
    file is checked with md5sum if it passes then it is unziped.
}
