% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tech_intensity.R
\name{get_tech_intensity}
\alias{get_tech_intensity}
\title{A function which creates a data frame of technology intensity for all
NACE coded industries in WIOD files.}
\usage{
get_tech_intensity()
}
\description{
A function which creates a data frame of technology intensity for all
NACE coded industries in WIOD files.
}
