% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_wiod.R
\name{get_wiod}
\alias{get_wiod}
\title{A function to donwload WIOD zip file from the project web page. The
downloaded file is the 2016 version which contains RData for each
year for the period 2000:2014. The integrity of the downloaded file
is checked with md5sum if it passes then it is unziped.}
\usage{
get_wiod()
}
\description{
A function to donwload WIOD zip file from the project web page. The
downloaded file is the 2016 version which contains RData for each
year for the period 2000:2014. The integrity of the downloaded file
is checked with md5sum if it passes then it is unziped.
}
\examples{
# Download the zip file to the default directory (wiod_original_data)
# from the default URL.

}
