% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_complementary.R
\name{getComplementary}
\alias{getComplementary}
\title{Generate a long table for VA, TXSP etc and other complementary values.}
\usage{
getComplementary(yearly.complementary, industry.RNr, value)
}
\arguments{
\item{yearly.complementary}{the last portion of the yearly WIOD data
containing complementary values.}

\item{industry.RNr}{the dataframe containing industry and its RNr
code to be used in column name}

\item{value}{one of the six different values; VA, II_fob, TXSP, EXP_adj, PURR, RURNR.}
}
\value{
data frame
}
\description{
Generate yearly long tables for six yearly different
    values for each economic sectors. These values are total
    intermediate consumption (II_fob), taxes less subsidies on
    products (TXSP), Cif/ fob adjustments on exports (EXP_adj),
    direct purchases abroad by residents (PURR), purchases on the
    domestic territory by non-residents (PURNR), value added at basic
    prices (VA).
}
