% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_country_panel.R
\name{getCountryWIOD}
\alias{getCountryWIOD}
\title{Create country based long tables and merge them to get a panel data set.}
\usage{
getCountryWIOD()
}
\description{
Create country based long tables and merge them them to
    create a panel data in csv form based on downloaded and prapared
    WIOD data. If there is no WIOD data downloaded, this function
    will donwload it, unzip it and prepare long tables.
}
\examples{
# Create country based panel data.
\dontrun{getCountryWIOD()}

}
