% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_long_tables.R
\name{getLongTables}
\alias{getLongTables}
\title{Generate long tables from the original WIOD zip file.}
\usage{
getLongTables(file.name)
}
\arguments{
\item{file.name}{yearly file name obtained from the download directory}
}
\description{
Call various functions to prepare long tables which are
    written on to the disk.
}
