% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tech_intensity.R
\name{getTechIntensity}
\alias{getTechIntensity}
\title{Create a data frame for technology intensity.}
\usage{
getTechIntensity(yearly.raw, isic)
}
\arguments{
\item{yearly.raw}{yearly raw data from the downloaded zip such as WIOT2011_October16_ROW.RData}

\item{isic}{parameter to be given for the aggregation of the industries.}
}
\description{
Create a data frame of technology intensity as defined
    by EUROSATAT for all NACE coded industries in WIOD files.
}
