% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_wiod_files.R
\name{list_files}
\alias{list_files}
\title{A function to list yearly, 15 WIOD files obtained after unzip of the
orginal downloaded file. All files are in the same directory.}
\usage{
list_files()
}
\description{
A function to list yearly, 15 WIOD files obtained after unzip of the
orginal downloaded file. All files are in the same directory.
}
