% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_network.R
\name{netCalc}
\alias{netCalc}
\title{Network calculation for a data frame with source - target - weight.}
\usage{
netCalc(wiod.net.df, ctry)
}
\arguments{
\item{wiod.net.df}{the long table data frame on which all network
calculations are done.}

\item{ctry}{control variable if the analysis is country based, 0 or
1.}
}
\description{
Calculate network scores of long files obtained from
    WIOD. The weight is filtered with 0.05 (for now).
}
