% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_network.R
\name{netCalcWrite}
\alias{netCalcWrite}
\title{Writing the calculated network file on disk.}
\usage{
netCalcWrite(wiod.long.file.name)
}
\arguments{
\item{wiod.long.file.name}{rda file containing the long network table
to read.}
}
\description{
Calculate network scores for all industries for each
    year is written on disk with wiod_network_score_YEAR.rda
}
