% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_network.R
\name{netWIOD}
\alias{netWIOD}
\title{Calculate network scores for all industries}
\usage{
netWIOD(file.dir = "./wiod_long_data")
}
\arguments{
\item{file.dir}{The directory into which all long tables obtained from
WIOD are written, Default: "./wiod_long_data"}
}
\description{
Calculate network scores for all industries for each
    year from the long tables created with getWIOD function.
}
