% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel_data.R
\name{panelWIOD}
\alias{panelWIOD}
\title{Merge all to create a panel data for analysis.}
\usage{
panelWIOD(isic)
}
\arguments{
\item{isic}{parameter to be given for the aggregation of the industries.}
}
\description{
Merge all long tables to create a panel data in csv form
    based on WIOD downloaded data.
}
\examples{
# Merge all long tables using default values.
\dontrun{panelWIOD(2)}

}
