% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_directories.R
\name{setDir}
\alias{setDir}
\title{Set directory names which store all files during data preparation}
\usage{
setDir(isic)
}
\arguments{
\item{isic}{parameter to be given for the aggregation of the industries.}
}
\value{
a list containing directories and final file name.
}
\description{
Set the names of all directories which store long tables
    created during data preparation according to the ISIC code
    given. If ISIC is 0, then industries in WIOD are not aggregated,
    1; aggregation at ISIC 1 digit, 2; aggregation in the
    manufacturing industry with respect to the Eurostat technology
    intensity classification and leaving all other sectors as they
    are.
}
