% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel_data.R
\name{bindFiles}
\alias{bindFiles}
\title{Bind all files to create a yearly panel data.}
\usage{
bindFiles(year, long.dir, net.dir, merge.dir, ctry)
}
\arguments{
\item{year}{between 2000 and 2014, WIOD years.}

\item{long.dir}{directory where the long tables are recorded, Default: "./wiod_long_data"}

\item{net.dir}{directory where the long tables on network
calculations are recorded, Default: "./wiod_network_data"}

\item{merge.dir}{directory where the long tables are all merged by
year basis, Default: "./yearly_merged_data"}

\item{ctry}{control variable if the process is based on countries}
}
\description{
Merge all long tables to create a panel data in csv form
    for each year.
}
