% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_wiod.R
\name{get_wiod}
\alias{get_wiod}
\title{A function to donwload WIOD zip file from the project web page.  The
downloaded file is the 2016 version which contains RData for each
year for the period 2000:2014. The integrity of the downloaded file
is checked with md5sum.}
\usage{
get_wiod(download_dir, data_URL)
}
\arguments{
\item{download_dir}{The directory into which the zip file is downloaded}

\item{data_URL}{The URL adress of the original WIOD data}
}
\description{
A function to donwload WIOD zip file from the project web page.  The
downloaded file is the 2016 version which contains RData for each
year for the period 2000:2014. The integrity of the downloaded file
is checked with md5sum.
}
\examples{
# Download to the default directory (wiod_original_data) and from
# the default URL.

# Download to the user defined directory

}
