% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_country_panel.R
\name{getCountryWIOD}
\alias{getCountryWIOD}
\title{Create country based long tables and merge them to get a panel data set.}
\usage{
getCountryWIOD(country.long.dir = "./wiod_long_data")
}
\arguments{
\item{country.long.dir}{directory where the long tables are recorded, Default: "./wiod_country_long_data"}
}
\description{
Create country based long tables and merge them them to
    create a panel data in csv form based on downloaded and prapared
    WIOD data.
}
\examples{
# Merge all long tables using default values.
\dontrun{getCountryWIOD()}

}
