% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_long_tables.R
\name{getLongTables}
\alias{getLongTables}
\title{Generate long tables from the original WIOD zip file.}
\usage{
getLongTables(file.name, dir.to.write, isic)
}
\arguments{
\item{file.name}{yearly file name obtained from the download directory}

\item{dir.to.write}{directory where all long files are recorded}

\item{isic}{parameter to be given for the aggregation of the industries.}
}
\description{
Call various functions to prepare long tables which are
    written on to the disk.
}
