% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_yearly_long_wiod.R
\name{getWIOD}
\alias{getWIOD}
\title{Call all functions to download and create yearly long tables.}
\usage{
getWIOD(download.dir = "./wiod_original_data")
}
\arguments{
\item{download.dir}{The directory into which the original WIOD zip
file is downloaded and unziped, Default: "./wiod_original_data"}
}
\description{
Call various functions to download, unzip, divide and
    create yearly long tables from the original WIOD file. All long
    tables are also saved.
}
\examples{
# Downloading the zip file to the default directory (wiod_original_data)
\dontrun{getWIOD(download.dir =  "./wiod_original_data")}

# Downloading the original zip file to the /user/defined/directory.
\dontrun{getWIOD("/user/defined/directory")}

}
