% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_calculations.R
\name{netCalcWrite}
\alias{netCalcWrite}
\title{Writing the calculated network file on disk.}
\usage{
netCalcWrite(wiod.long.file.name, net.data.dir, ctry)
}
\arguments{
\item{wiod.long.file.name}{rda file containing the long network table
to read.}

\item{net.data.dir}{directory where all network calculation files
are stored.}

\item{ctry}{control variable if the analysis is country based, 0 or
1.}
}
\description{
Calculate network scores for all industries for each
    year is written on disk with wiod_network_score_YEAR.rda
}
