% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_calculations.R
\name{netWIOD}
\alias{netWIOD}
\title{Calculate network scores for all industries}
\usage{
netWIOD(isic)
}
\arguments{
\item{isic}{parameter to be given for the aggregation of the industries.}
}
\description{
Calculate network scores for all industries for each
    year from the long tables created with getWIOD function.
}
